# Faster stringer implementation for Address