## Summary

Just make it faster. Nothing exciting going here.

## Benchmark Plan
Address.String():  550 ns / op -> 350 ns / op
UnmarshalChecksumAddress():  1116 ns / op -> 1001 ns / op